・クラス名変更 ・MainForm.csのメソッド分離（→MainForm.Utils.cs） ・目次記号混入問題解決
