+++ src/MJS_fileJoin_src/MJS_fileJoin/MainForm.Config.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MJS_fileJoin
+{
+    public partial class MainForm
+    {
+
+    }
+}
