+++ src/MJS_fileJoin_src/MJS_fileJoin/Program.cs
@@ -16,7 +16,7 @@ namespace MJS_fileJoin
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new MainForm());
         }
     }
 }
