+++ src/MJS_fileJoin_src/MJS_fileJoin/define.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace MJS_fileJoin
 {
-    partial class Form1
+    partial class MainForm
     {
         //const string searchJs = @"var searchWords = $('♪');" + "\n" + @"$(function(){" + "\n" + @"  $("".wSearchField"").each(function() {" + "\n" + @"    $(this).off();" + "\n" + @"  });" + "\n" + @"  $(document).on(""keyup"", "".wSearchField"", function(){" + "\n" + @"    if($(this).val() == """")" + "\n" + @"    {" + "\n" + @"      $("".wSearchResultItemsBlock"").html("""");" + "\n" + @"      $("".wSearchResultsEnd"").addClass(""rh-hide"");" + "\n" + @"      $("".wSearchResultsEnd"").attr(""hidden"", """");" + "\n" + @"      $(""#searchMsg"").html("""");" + "\n" + @"      $(""iframe.topic"").contents().find("".keyword"").each(function() {" + "\n" + @"        for(var i = 0; i < $(this)[0].childNodes.length; i ++)" + "\n" + @"        {" + "\n" + @"          this.parentNode.insertBefore($(this)[0].childNodes[i], this)" + "\n" + @"        }" + "\n" + @"        $(this).remove();" + "\n" + @"      });" + "\n" + @"    }" + "\n" + @"    else" + "\n" + @"    {" + "\n" + @"      var findItems = searchWords.find("".search_word:contains('""+$(this).val()+""')"");" + "\n" + @"      if(findItems.length != 0)" + "\n" + @"      {" + "\n" + @"        $("".wSearchResultsEnd"").removeClass(""rh-hide"");" + "\n" + @"        $("".wSearchResultsEnd"").removeAttr(""hidden"");" + "\n" + @"        $("".wSearchResultItemsBlock"").html("""");" + "\n" + @"        findItems.each(function() {" + "\n" + @"          var searchWord = $(this).parent().find("".search_word"").text();" + "\n" + @"          if(searchWord.length > 90)" + "\n" + @"          {" + "\n" + @"            searchWord = searchWord.substring(0, 90) + "" ..."";" + "\n" + @"          }" + "\n" + @"          $("".wSearchResultItemsBlock"").append($(""<div class='wSearchResultItem'><a class='nolink' href='./""+$(this).parent().attr(""id"")+"".html?rhhlterm=%E3%80%82&rhsyns=%20'><div class='wSearchResultTitle'>""+$(this).parent().find("".search_title"").html()+""</div></a><div class='wSearchContent'><span class='wSearchContext'>""+searchWord+""</span></div></div>""));" + "\n" + @"        });" + "\n" + @"        $(""iframe.topic"").contents().find("".keyword"").each(function() {" + "\n" + @"          for(var i = 0; i < $(this)[0].childNodes.length; i ++)" + "\n" + @"          {" + "\n" + @"            this.parentNode.insertBefore($(this)[0].childNodes[i], this)" + "\n" + @"          }" + "\n" + @"          $(this).remove();" + "\n" + @"        });" + "\n" + @"        var reg = new RegExp(""(""+$("".wSearchField"", document).val().replace(""&"",""&amp;"").replace(""<"",""&lt;"")+"")(?=[^<>]*<)"", ""gm"");" + "\n" + @"        $(""iframe.topic"").contents().find(""body"").html($(""iframe.topic"").contents().find(""body"").html().replace(reg, ""<font class='keyword' style='color:rgb(0, 0, 0); background-color:rgb(252, 255, 0);'>$1</font>""));" + "\n" + @"      }" + "\n" + @"      else" + "\n" + @"      {" + "\n" + @"        $(""iframe.topic"").contents().find("".keyword"").each(function() {" + "\n" + @"          for(var i = 0; i < $(this)[0].childNodes.length; i ++)" + "\n" + @"          {" + "\n" + @"            this.parentNode.insertBefore($(this)[0].childNodes[i], this)" + "\n" + @"          }" + "\n" + @"          $(this).remove();" + "\n" + @"        });" + "\n" + @"        $("".wSearchResultsEnd"").addClass(""rh-hide"");" + "\n" + @"        $("".wSearchResultsEnd"").attr(""hidden"", """");" + "\n" + @"        $("".wSearchResultItemsBlock"").html("""");" + "\n" + "\n" + @"      }" + "\n" + @"    }" + "\n" + @"  });" + "\n" + @"  $(""iframe.topic"").on(""load"", function(){" + "\n" + @"    if($("".search-input"", document).is("":not(.rh-hide)"") && ($("".wSearchField"", document).val() != """"))" + "\n" + @"    {" + "\n" + @"      var reg = new RegExp(""(""+$("".wSearchField"", document).val().replace(""&"",""&amp;"").replace(""<"",""&lt;"")+"")(?=[^<>]*<)"", ""gm"");" + "\n" + @"      $(""iframe.topic"").contents().find(""body"").html($(""iframe.topic"").contents().find(""body"").html().replace(reg, ""<font class='keyword' style='color:rgb(0, 0, 0); background-color:rgb(252, 255, 0);'>$1</font>""));" + "\n" + @"    }" + "\n" + @"  });" + "\n" + @"});" + "\n";
         const string searchJs = @"var searchWords = $('♪');" + "\n" + @"var wide = Array(""０"",""１"",""２"",""３"",""４"",""５"",""６"",""７"",""８"",""９"",""Ａ"",""Ｂ"",""Ｃ"",""Ｄ"",""Ｅ"",""Ｆ"",""Ｇ"",""Ｈ"",""Ｉ"",""Ｊ"",""Ｋ"",""Ｌ"",""Ｍ"",""Ｎ"",""Ｏ"",""Ｐ"",""Ｑ"",""Ｒ"",""Ｓ"",""Ｔ"",""Ｕ"",""Ｖ"",""Ｗ"",""Ｘ"",""Ｙ"",""Ｚ"",""ａ"",""ｂ"",""ｃ"",""ｄ"",""ｅ"",""ｆ"",""ｇ"",""ｈ"",""ｉ"",""ｊ"",""ｋ"",""ｌ"",""ｍ"",""ｎ"",""ｏ"",""ｐ"",""ｑ"",""ｒ"",""ｓ"",""ｔ"",""ｕ"",""ｖ"",""ｗ"",""ｘ"",""ｙ"",""ｚ"",""ガ"",""ギ"",""グ"",""ゲ"",""ゴ"",""ザ"",""ジ"",""ズ"",""ゼ"",""ゾ"",""ダ"",""ヂ"",""ヅ"",""デ"",""ド"",""バ"",""ビ"",""ブ"",""ベ"",""ボ"",""パ"",""ピ"",""プ"",""ペ"",""ポ"",""。"",""「"",""」"",""、"",""ヲ"",""ァ"",""ィ"",""ゥ"",""ェ"",""ォ"",""ャ"",""ュ"",""ョ"",""ッ"",""ー"",""ア"",""イ"",""ウ"",""エ"",""オ"",""カ"",""キ"",""ク"",""ケ"",""コ"",""サ"",""シ"",""ス"",""セ"",""ソ"",""タ"",""チ"",""ツ"",""テ"",""ト"",""ナ"",""ニ"",""ヌ"",""ネ"",""ノ"",""ハ"",""ヒ"",""フ"",""ヘ"",""ホ"",""マ"",""ミ"",""ム"",""メ"",""モ"",""ヤ"",""ユ"",""ヨ"",""ラ"",""リ"",""ル"",""レ"",""ロ"",""ワ"",""ン"");" + "\n" + @"var narrow = Array(""0"",""1"",""2"",""3"",""4"",""5"",""6"",""7"",""8"",""9"",""a"",""b"",""c"",""d"",""e"",""f"",""g"",""h"",""i"",""j"",""k"",""l"",""m"",""n"",""o"",""p"",""q"",""r"",""s"",""t"",""u"",""v"",""w"",""x"",""y"",""z"",""a"",""b"",""c"",""d"",""e"",""f"",""g"",""h"",""i"",""j"",""k"",""l"",""m"",""n"",""o"",""p"",""q"",""r"",""s"",""t"",""u"",""v"",""w"",""x"",""y"",""z"",""ｶﾞ"",""ｷﾞ"",""ｸﾞ"",""ｹﾞ"",""ｺﾞ"",""ｻﾞ"",""ｼﾞ"",""ｽﾞ"",""ｾﾞ"",""ｿﾞ"",""ﾀﾞ"",""ﾁﾞ"",""ﾂﾞ"",""ﾃﾞ"",""ﾄﾞ"",""ﾊﾞ"",""ﾋﾞ"",""ﾌﾞ"",""ﾍﾞ"",""ﾎﾞ"",""ﾊﾟ"",""ﾋﾟ"",""ﾌﾟ"",""ﾍﾟ"",""ﾎﾟ"",""｡"",""｢"",""｣"",""､"",""ｦ"",""ｧ"",""ｨ"",""ｩ"",""ｪ"",""ｫ"",""ｬ"",""ｭ"",""ｮ"",""ｯ"",""ｰ"",""ｱ"",""ｲ"",""ｳ"",""ｴ"",""ｵ"",""ｶ"",""ｷ"",""ｸ"",""ｹ"",""ｺ"",""ｻ"",""ｼ"",""ｽ"",""ｾ"",""ｿ"",""ﾀ"",""ﾁ"",""ﾂ"",""ﾃ"",""ﾄ"",""ﾅ"",""ﾆ"",""ﾇ"",""ﾈ"",""ﾉ"",""ﾊ"",""ﾋ"",""ﾌ"",""ﾍ"",""ﾎ"",""ﾏ"",""ﾐ"",""ﾑ"",""ﾒ"",""ﾓ"",""ﾔ"",""ﾕ"",""ﾖ"",""ﾗ"",""ﾘ"",""ﾙ"",""ﾚ"",""ﾛ"",""ﾜ"",""ﾝ"");" + "\n" + @"var hilight = Array(""(?:０|0)"",""(?:１|1)"",""(?:２|2)"",""(?:３|3)"",""(?:４|4)"",""(?:５|5)"",""(?:６|6)"",""(?:７|7)"",""(?:８|8)"",""(?:９|9)"",""(?:Ａ|A|ａ|a)"",""(?:Ｂ|B|ｂ|b)"",""(?:Ｃ|C|ｃ|c)"",""(?:Ｄ|D|ｄ|d)"",""(?:Ｅ|E|ｅ|e)"",""(?:Ｆ|F|ｆ|f)"",""(?:Ｇ|G|ｇ|g)"",""(?:Ｈ|H|ｈ|h)"",""(?:Ｉ|I|ｉ|i)"",""(?:Ｊ|J|ｊ|j)"",""(?:Ｋ|K|ｋ|k)"",""(?:Ｌ|L|ｌ|l)"",""(?:Ｍ|M|ｍ|m)"",""(?:Ｎ|N|ｎ|n)"",""(?:Ｏ|O|ｏ|o)"",""(?:Ｐ|P|ｐ|p)"",""(?:Ｑ|Q|ｑ|q)"",""(?:Ｒ|R|ｒ|r)"",""(?:Ｓ|S|ｓ|s)"",""(?:Ｔ|T|ｔ|t)"",""(?:Ｕ|U|ｕ|u)"",""(?:Ｖ|V|ｖ|v)"",""(?:Ｗ|W|ｗ|w)"",""(?:Ｘ|X|ｘ|x)"",""(?:Ｙ|Y|ｙ|y)"",""(?:Ｚ|Z|ｚ|z)"",""(?:ガ|ｶﾞ)"",""(?:ギ|ｷﾞ)"",""(?:グ|ｸﾞ)"",""(?:ゲ|ｹﾞ)"",""(?:ゴ|ｺﾞ)"",""(?:ザ|ｻﾞ)"",""(?:ジ|ｼﾞ)"",""(?:ズ|ｽﾞ)"",""(?:ゼ|ｾﾞ)"",""(?:ゾ|ｿﾞ)"",""(?:ダ|ﾀﾞ)"",""(?:ヂ|ﾁﾞ)"",""(?:ヅ|ﾂﾞ)"",""(?:デ|ﾃﾞ)"",""(?:ド|ﾄﾞ)"",""(?:バ|ﾊﾞ)"",""(?:ビ|ﾋﾞ)"",""(?:ブ|ﾌﾞ)"",""(?:ベ|ﾍﾞ)"",""(?:ボ|ﾎﾞ)"",""(?:パ|ﾊﾟ)"",""(?:ピ|ﾋﾟ)"",""(?:プ|ﾌﾟ)"",""(?:ペ|ﾍﾟ)"",""(?:ポ|ﾎﾟ)"",""(?:。|｡)"",""(?:「|｢)"",""(?:」|｣)"",""(?:、|､)"",""(?:ヲ|ｦ)"",""(?:ァ|ｧ)"",""(?:ィ|ｨ)"",""(?:ゥ|ｩ)"",""(?:ェ|ｪ)"",""(?:ォ|ｫ)"",""(?:ャ|ｬ)"",""(?:ュ|ｭ)"",""(?:ョ|ｮ)"",""(?:ッ|ｯ)"",""(?:ー|ｰ)"",""(?:ア|ｱ)"",""(?:イ|ｲ)"",""(?:ウ|ｳ)"",""(?:エ|ｴ)"",""(?:オ|ｵ)"",""(?:カ|ｶ)"",""(?:キ|ｷ)"",""(?:ク|ｸ)"",""(?:ケ|ｹ)"",""(?:コ|ｺ)"",""(?:サ|ｻ)"",""(?:シ|ｼ)"",""(?:ス|ｽ)"",""(?:セ|ｾ)"",""(?:ソ|ｿ)"",""(?:タ|ﾀ)"",""(?:チ|ﾁ)"",""(?:ツ|ﾂ)"",""(?:テ|ﾃ)"",""(?:ト|ﾄ)"",""(?:ナ|ﾅ)"",""(?:ニ|ﾆ)"",""(?:ヌ|ﾇ)"",""(?:ネ|ﾈ)"",""(?:ノ|ﾉ)"",""(?:ハ|ﾊ)"",""(?:ヒ|ﾋ)"",""(?:フ|ﾌ)"",""(?:ヘ|ﾍ)"",""(?:ホ|ﾎ)"",""(?:マ|ﾏ)"",""(?:ミ|ﾐ)"",""(?:ム|ﾑ)"",""(?:メ|ﾒ)"",""(?:モ|ﾓ)"",""(?:ヤ|ﾔ)"",""(?:ユ|ﾕ)"",""(?:ヨ|ﾖ)"",""(?:ラ|ﾗ)"",""(?:リ|ﾘ)"",""(?:ル|ﾙ)"",""(?:レ|ﾚ)"",""(?:ロ|ﾛ)"",""(?:ワ|ﾜ)"",""(?:ン|ﾝ)"");" + "\n" + @"function selectorEscape(val){" + "\n" + @"  return val.replace(/[-\/\\^$*+?.()|[\]{}\!]/g, '\\$&');" + "\n" + @"}" + "\n" + @"" + "\n" + @"$(function(){" + "\n" + @"  $(document).on(""click"", ""ul.toc li.book"", function() {" + "\n" + @"    if($(this).children(""a[href='#'],a[href='javascript:void 0;']"").length == 0)" + "\n" + @"    {" + "\n" + @"      $(this).children(""a"").each(function(){" + "\n" + @"        location.href=location.href.replace(location.hash,"""")+""#t=""+$(this).attr(""href"");" + "\n" + @"      });" + "\n" + @"    }" + "\n" + @"  });" + "\n" + @"  $("".wSearchField"").each(function() {" + "\n" + @"    $(this).off();" + "\n" + @"  });" + "\n" + @"  $(document).on(""keyup"", "".wSearchField"", function(){" + "\n" + @"    if($(this).val() == """")" + "\n" + @"    {" + "\n" + @"      $("".wSearchResultItemsBlock"").html("""");" + "\n" + @"      $("".wSearchResultsEnd"").addClass(""rh-hide"");" + "\n" + @"      $("".wSearchResultsEnd"").attr(""hidden"", """");" + "\n" + @"      $(""#searchMsg"").html(""2つ以上の語句を入力して検索する場合は、スペース（空白）で区切ります。"");" + "\n" + @"      $(""iframe.topic"").contents().find("".keyword"").each(function() {" + "\n" + @"        for(var i = 0; i < $(this)[0].childNodes.length; i ++)" + "\n" + @"        {" + "\n" + @"          this.parentNode.insertBefore($(this)[0].childNodes[i], this)" + "\n" + @"        }" + "\n" + @"        $(this).remove();" + "\n" + @"      });" + "\n" + @"    }" + "\n" + @"    else" + "\n" + @"    {" + "\n" + @"      $(""#searchMsg"").html("""");" + "\n" + @"      var searchWordTmp = $(this).val().replace(/(.*?)(?:　| )+(.*?)/g, ""$1 $2"").trim().toLowerCase();" + "\n" + @"      for(i = 0; i < wide.length; i ++)" + "\n" + @"      {" + "\n" + @"        searchWordTmp = searchWordTmp.split(wide[i]).join(narrow[i]);" + "\n" + @"      }" + "\n" + @"      var searchWord = searchWordTmp.split("" "");" + "\n" + @"      var searchQuery = """";" + "\n" + @"      for(i = 0; i < searchWord.length; i ++)" + "\n" + @"      {" + "\n" + @"        searchQuery += "":contains("" + searchWord[i] + "")"";" + "\n" + @"      }" + "\n" + @"      " + "\n" + @"      var findItems = searchWords.find("".search_word""+searchQuery);" + "\n" + @"      if(findItems.length != 0)" + "\n" + @"      {" + "\n" + @"        $("".wSearchResultsEnd"").removeClass(""rh-hide"");" + "\n" + @"        $("".wSearchResultsEnd"").removeAttr(""hidden"");" + "\n" + @"        $("".wSearchResultItemsBlock"").html("""");" + "\n" + @"        findItems.each(function() {" + "\n" + @"          var displayText = $(this).parent().find("".displayText"").text();" + "\n" + @"          $("".wSearchResultItemsBlock"").append($(""<div class='wSearchResultItem'><a class='nolink' href='./""+$(this).parent().attr(""id"")+"".html'><div class='wSearchResultTitle'>""+$(this).parent().find("".search_title"").html()+""</div></a><div class='wSearchContent'><span class='wSearchContext'>""+displayText+""</span></div></div>""));" + "\n" + @"        });" + "\n" + @"        $(""iframe.topic"").contents().find("".keyword"").each(function() {" + "\n" + @"          for(var i = 0; i < $(this)[0].childNodes.length; i ++)" + "\n" + @"          {" + "\n" + @"            this.parentNode.insertBefore($(this)[0].childNodes[i], this)" + "\n" + @"          }" + "\n" + @"          $(this).remove();" + "\n" + @"        });" + "\n" + @"        for(i = 0; i < searchWord.length; i ++)" + "\n" + @"        {" + "\n" + @"        	searchWord[i] = selectorEscape(searchWord[i].replace("">"", ""&gt;"").replace(""<"", ""&lt;""));" + "\n" + @"        }" + "\n" + @"        " + "\n" + @"        var hilightWord = searchWord.join(""|"");" + "\n" + @"        for(i = 0; i < hilight.length; i ++)" + "\n" + @"        {" + "\n" + @"	        var reg = new RegExp(hilight[i], ""gm"");" + "\n" + @"        	hilightWord = hilightWord.replace(reg, hilight[i]);" + "\n" + @"        }" + "\n" + @"        var reg = new RegExp(""(""+hilightWord+"")(?=[^<>]*<)"", ""gm"");" + "\n" + @"        var regnbsp = new RegExp(""&nbsp;(?=[^<>]*<)"", ""gm"");" + "\n" + @"        var reggt = new RegExp(""&gt;(?=[^<>]*<)"", ""gm"");" + "\n" + @"        var reglt = new RegExp(""&lt;(?=[^<>]*<)"", ""gm"");" + "\n" + @"        var regquot = new RegExp(""&quot;(?=[^<>]*<)"", ""gm"");" + "\n" + @"        var regamp = new RegExp(""&amp;(?=[^<>]*<)"", ""gm"");" + "\n" + @"        $(""iframe.topic"").contents().find(""body"").html($(""iframe.topic"").contents().find(""body"").html().replace(regnbsp, ""　"").replace(reggt, "">"").replace(reglt, ""<"").replace(regquot, '""').replace(regamp, ""&"").replace(reg, ""<font class='keyword' style='color:rgb(0, 0, 0); background-color:rgb(252, 255, 0);'>$1</font>""));" + "\n" + @"      }" + "\n" + @"      else" + "\n" + @"      {" + "\n" + @"        $(""iframe.topic"").contents().find("".keyword"").each(function() {" + "\n" + @"          for(var i = 0; i < $(this)[0].childNodes.length; i ++)" + "\n" + @"          {" + "\n" + @"            this.parentNode.insertBefore($(this)[0].childNodes[i], this)" + "\n" + @"          }" + "\n" + @"          $(this).remove();" + "\n" + @"        });" + "\n" + @"        $("".wSearchResultsEnd"").addClass(""rh-hide"");" + "\n" + @"        $("".wSearchResultsEnd"").attr(""hidden"", """");" + "\n" + @"        $("".wSearchResultItemsBlock"").html("""");" + "\n" + @"        displayText = ""検索条件に一致するトピックはありません。"";" + "\n" + @"        $("".wSearchResultItemsBlock"").append($(""<div class='wSearchResultItem'><div class='wSearchContent'><span class='wSearchContext'>""+displayText+""</span></div></div>""));" + "\n" + @"        //this.parentNode.insertBefore($(this)[0].childNodes[0], this);" + "\n" + @"      }" + "\n" + @"    }" + "\n" + @"  });" + "\n" + @"  $(""iframe.topic"").on(""load"", function(){" + "\n" + @"    if($("".search-input"", document).is("":not(.rh-hide)"") && ($("".wSearchField"", document).val() != """"))" + "\n" + @"    {" + "\n" + @"      var searchWordTmp = $("".wSearchField"", document).val().split(""　"").join("" "").trim();" + "\n" + @"      searchWordTmp = searchWordTmp.split(""  "").join("" "");" + "\n" + @"      for(i = 0; i < wide.length; i ++)" + "\n" + @"      {" + "\n" + @"        searchWordTmp = searchWordTmp.replace(wide[i], narrow[i]);" + "\n" + @"      }" + "\n" + @"      var searchWord = searchWordTmp.split("" "");" + "\n" + @"      for(i = 0; i < searchWord.length; i ++)" + "\n" + @"      {" + "\n" + @"        searchWord[i] = selectorEscape(searchWord[i].replace("">"", ""&gt;"").replace(""<"", ""&lt;""));" + "\n" + @"      }" + "\n" + @"      var hilightWord = searchWord.join(""|"");" + "\n" + @"      for(i = 0; i < hilight.length; i ++)" + "\n" + @"      {" + "\n" + @"        var reg = new RegExp(hilight[i], ""gm"");" + "\n" + @"        hilightWord = hilightWord.replace(reg, hilight[i]);" + "\n" + @"      }" + "\n" + @"" + "\n" + @"      var reg = new RegExp(""(""+hilightWord+"")(?=[^<>]*<)"", ""gm"");" + "\n" + @"      var regnbsp = new RegExp(""&nbsp;(?=[^<>]*<)"", ""gm"");" + "\n" + @"      var reggt = new RegExp(""&gt;(?=[^<>]*<)"", ""gm"");" + "\n" + @"      var reglt = new RegExp(""&lt;(?=[^<>]*<)"", ""gm"");" + "\n" + @"      var regquot = new RegExp(""&quot;(?=[^<>]*<)"", ""gm"");" + "\n" + @"      var regamp = new RegExp(""&amp;(?=[^<>]*<)"", ""gm"");" + "\n" + @"        $(""iframe.topic"").contents().find(""body"").html($(""iframe.topic"").contents().find(""body"").html().replace(regnbsp, ""　"").replace(reggt, "">"").replace(reglt, ""<"").replace(regquot, '""').replace(regamp, ""&"").replace(reg, ""<font class='keyword' style='color:rgb(0, 0, 0); background-color:rgb(252, 255, 0);'>$1</font>""));" + "\n" + @"    }" + "\n" + @"  });" + "\n" + @"});" + "\n";
